// x-pack/platform/plugins/shared/aiops/public/embeddables/pattern_analysis/types.d.ts
import type { DefaultEmbeddableApi } from '@kbn/embeddable-plugin/public';
import type { HasEditCapabilities, PublishesDataViews, PublishesTimeRange, PublishingSubject } from '@kbn/presentation-publishing';
import type { PatternAnalysisEmbeddableState } from '../../../common/embeddables/pattern_analysis/types';
export interface PatternAnalysisComponentApi {
    dataViewId: PublishingSubject<PatternAnalysisEmbeddableState['dataViewId']>;
    fieldName: PublishingSubject<PatternAnalysisEmbeddableState['fieldName']>;
    minimumTimeRangeOption: PublishingSubject<PatternAnalysisEmbeddableState['minimumTimeRangeOption']>;
    randomSamplerMode: PublishingSubject<PatternAnalysisEmbeddableState['randomSamplerMode']>;
    randomSamplerProbability: PublishingSubject<PatternAnalysisEmbeddableState['randomSamplerProbability']>;
    updateUserInput: (update: PatternAnalysisEmbeddableState) => void;
}
export type PatternAnalysisEmbeddableApi = DefaultEmbeddableApi<PatternAnalysisEmbeddableState> & HasEditCapabilities & PublishesDataViews & PublishesTimeRange & PatternAnalysisComponentApi;